+++ takei/blog/app/Http/Controllers/PostController.php
@@ -20,7 +20,7 @@ class PostController extends Controller
     public function index()
     {
         $posts = Post::fetchPosts();
-        return Inertia::render('Posts/Index',['posts' => $posts]);
+        return Inertia::render('Posts/List',['posts' => $posts]);
     }
 
     /**
